Airflow/requirements.txt
apache-airflow-providers-apache-spark
minio
pyspark